Refactor Scala 3 macros
